rename: 폴더 구조 변경으로 인한 파일, 폴더명 변경 #11
